Make text representation more intuitive
As discussed in issue #99, the text() method now renders whitespace
more like a browser and uses the original whitespace of the html but
strips consecutive whitespace.

An exception is made for query results since they will never have any
separating whitespace. In this case a space will be added between each
element to improve readability.